+++ src/soplex.h
@@ -1500,10 +1500,10 @@ public:
    bool areLPsInSync(const bool checkVecVals = true, const bool checkMatVals = false, const bool quiet = false) const;
 
    /// set the random seeds of the solver instance
-   void setRandomSeed(unsigned long seed);
+   void setRandomSeed(unsigned int seed);
 
    /// returns the current random seed of the solver instance
-   unsigned long randomSeed() const;
+   unsigned int randomSeed() const;
 
    //@}
